refact : 비밀 캡슐 생성 기능 위치 이동
