+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/generic_capsule/api/CapsuleApiController.java
@@ -1,11 +1,14 @@
 package site.timecapsulearchive.core.domain.capsule.generic_capsule.api;
 
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,6 +17,8 @@ import site.timecapsulearchive.core.domain.capsule.generic_capsule.data.dto.Coor
 import site.timecapsulearchive.core.domain.capsule.generic_capsule.data.response.CapsuleOpenedResponse;
 import site.timecapsulearchive.core.domain.capsule.generic_capsule.data.response.ImagesPageResponse;
 import site.timecapsulearchive.core.domain.capsule.generic_capsule.data.response.NearbyCapsuleResponse;
+import site.timecapsulearchive.core.domain.capsule.mapper.CapsuleMapper;
+import site.timecapsulearchive.core.domain.capsule.secret_capsule.data.reqeust.CapsuleCreateRequest;
 import site.timecapsulearchive.core.domain.capsule.service.CapsuleService;
 import site.timecapsulearchive.core.global.common.response.ApiSpec;
 import site.timecapsulearchive.core.global.common.response.SuccessCode;
@@ -24,6 +29,7 @@ import site.timecapsulearchive.core.global.common.response.SuccessCode;
 public class CapsuleApiController implements CapsuleApi {
 
     private final CapsuleService capsuleService;
+    private final CapsuleMapper capsuleMapper;
 
     @GetMapping(value = "/images", produces = {"application/json"})
     @Override
@@ -63,6 +69,42 @@ public class CapsuleApiController implements CapsuleApi {
             ApiSpec.success(
                 SuccessCode.SUCCESS,
                 capsuleService.updateCapsuleOpened(memberId, capsuleId)
+            )
+        );
+    }
+
+
+    @PostMapping(value = "/secret", consumes = {"application/json"})
+    @Override
+    public ResponseEntity<ApiSpec<String>> createSecretCapsule(
+        @AuthenticationPrincipal final Long memberId,
+        @Valid @RequestBody final CapsuleCreateRequest request
+    ) {
+        capsuleService.saveCapsule(
+            memberId,
+            capsuleMapper.CapsuleCreateRequestToDto(request)
+        );
+
+        return ResponseEntity.ok(
+            ApiSpec.empty(
+                SuccessCode.SUCCESS
+            )
+        );
+    }
+    @PostMapping(value = "/public", consumes = {"application/json"})
+    @Override
+    public ResponseEntity<ApiSpec<String>> createPublicCapsule(
+        @AuthenticationPrincipal final Long memberId,
+        @Valid @RequestBody final CapsuleCreateRequest request) {
+
+        capsuleService.saveCapsule(
+            memberId,
+            capsuleMapper.CapsuleCreateRequestToDto(request)
+        );
+
+        return ResponseEntity.ok(
+            ApiSpec.empty(
+                SuccessCode.SUCCESS
             )
         );
     }
