+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/service/SecretCapsuleService.java
@@ -8,7 +8,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import site.timecapsulearchive.core.domain.capsule.entity.Capsule;
 import site.timecapsulearchive.core.domain.capsule.exception.CapsuleNotFondException;
 import site.timecapsulearchive.core.domain.capsule.mapper.CapsuleMapper;
 import site.timecapsulearchive.core.domain.capsule.mapper.ImageMapper;
@@ -18,16 +17,11 @@ import site.timecapsulearchive.core.domain.capsule.repository.CapsuleRepository;
 import site.timecapsulearchive.core.domain.capsule.repository.ImageQueryRepository;
 import site.timecapsulearchive.core.domain.capsule.repository.VideoQueryRepository;
 import site.timecapsulearchive.core.domain.capsule.secret_capsule.data.dto.MySecreteCapsuleDto;
-import site.timecapsulearchive.core.domain.capsule.secret_capsule.data.dto.SecretCapsuleCreateRequestDto;
 import site.timecapsulearchive.core.domain.capsule.secret_capsule.data.dto.SecretCapsuleDetailDto;
 import site.timecapsulearchive.core.domain.capsule.secret_capsule.data.dto.SecretCapsuleSummaryDto;
 import site.timecapsulearchive.core.domain.capsule.secret_capsule.data.response.MySecretCapsuleSliceResponse;
 import site.timecapsulearchive.core.domain.capsule.secret_capsule.data.response.SecretCapsuleDetailResponse;
-import site.timecapsulearchive.core.domain.capsuleskin.entity.CapsuleSkin;
-import site.timecapsulearchive.core.domain.capsuleskin.exception.CapsuleSkinNotFoundException;
 import site.timecapsulearchive.core.domain.capsuleskin.repository.CapsuleSkinRepository;
-import site.timecapsulearchive.core.domain.member.entity.Member;
-import site.timecapsulearchive.core.domain.member.exception.MemberNotFoundException;
 import site.timecapsulearchive.core.domain.member.repository.MemberRepository;
 import site.timecapsulearchive.core.infra.s3.data.dto.S3PreSignedUrlDto;
 import site.timecapsulearchive.core.infra.s3.data.request.S3PreSignedUrlRequestDto;
@@ -135,43 +129,6 @@ public class SecretCapsuleService {
         }
 
         return !dto.isOpened() || dto.dueDate().isAfter(ZonedDateTime.now(ZoneOffset.UTC));
-    }
-
-    /**
-     * 멤버 아이디와 캡슐 생성 포맷을 받아서 캡슐을 생성한다.
-     *
-     * @param memberId 캡슐을 생성할 멤버 아이디
-     * @param dto      캡슐 생성 요청 포맷
-     */
-    @Transactional
-    public void saveCapsule(final Long memberId, final SecretCapsuleCreateRequestDto dto) {
-        final Member foundMember = memberRepository.findMemberById(memberId)
-            .orElseThrow(MemberNotFoundException::new);
-
-        final CapsuleSkin foundCapsuleSkin = capsuleSkinRepository
-            .findCapsuleSkinById(dto.capsuleSkinId())
-            .orElseThrow(CapsuleSkinNotFoundException::new);
-
-        final Capsule capsule = capsuleMapper.requestDtoToEntity(dto, foundMember,
-            foundCapsuleSkin);
-
-        capsuleRepository.save(capsule);
-
-        if (isNotEmpty(dto.imageNames())) {
-            imageQueryRepository.bulkSave(
-                imageMapper.toEntity(capsule, foundMember, dto.directory(), dto.imageNames())
-            );
-        }
-
-        if (isNotEmpty(dto.videoNames())) {
-            videoQueryRepository.bulkSave(
-                videoMapper.toEntity(capsule, foundMember, dto.directory(), dto.videoNames())
-            );
-        }
-    }
-
-    private boolean isNotEmpty(List<String> fileNames) {
-        return fileNames != null && !fileNames.isEmpty();
     }
 }
